have ls return uint instead of long

+++ Payload_Type/athena/athena/agent_code/Athena.Models/Responses/FileBrowserResponseResult.cs
@@ -29,8 +29,8 @@ namespace Athena.Models.Responses
         public string name { get; set; }
         public string? parent_path { get; set; }
         public bool success { get; set; }
-        public long access_time { get; set; }
-        public long modify_time { get; set; }
+        public uint access_time { get; set; }
+        public uint modify_time { get; set; }
         public long size { get; set; }
         public bool update_deleted { get; set; }
         public List<FileBrowserFile> files { get; set; }
@@ -41,8 +41,8 @@ namespace Athena.Models.Responses
         public bool is_file { get; set; }
         public Dictionary<string, string> permissions { get; set; }
         public string name { get; set; }
-        public long access_time { get; set; }
-        public long modify_time { get; set; }
+        public uint access_time { get; set; }
+        public uint modify_time { get; set; }
         public long size { get; set; }
     }
 }
